Fix gorm syntax error in IN clause

+++ internal/domain/model/energy_consumption.model.go
@@ -8,7 +8,7 @@ import (
 
 type EnergyConsumption struct {
 	Id          uuid.UUID `gorm:"type:uuid;primaryKey;default:gen_random_uuid()"`
-	DeviceId    uint      `gorm:"not null"`
-	Consumption float64   `gorm:"not null"`
+	DeviceId    uint8      `gorm:"not null"`
+	Consumption float32   `gorm:"not null"`
 	CreatedAt   time.Time `gorm:"autoCreateTime"`
 }
